Support an extra androidKey in KMT_CLICK and KMT_CLICK_TWICE
Using Android key "W" to move forward in PUBG moves the character
slowly (walk). To run with "W", I would configure "W" as KMT_CLICK_TWICE
on the Auto-run icon, and the functionality to trigger an Android key "W"
is still needed in the case of driving a vehicle.

An alternative approach is to configure Shift as KMT_CLICK_TWICE on
Auto-run. This would enable the "walk" functionality which was not
possible. However this would conflict with "boost" when driving since
the key for "boost" needs to be KMT_CLICK (not TWICE).

Users should be able to configure the keymap in whichever way they want.